Fix: send the errors on the err channles unblocking way
On the way, which can be unblocked by context. Otherwise, if
- we invoke multiple goroutines which are doing I/O
- they all run into I/O err
- they all submit the err via error channel back
only the first error gets consumed, all other goroutins are blocked and waiting for the error to be consumed.
Another side negative thing, nobody closes the channel on the clean way. We should consider this handling as another pattern.

Signed-off-by: Artem Sidorenko <artem@posteo.de>